Update bitops.h, primesieve.c
Update makeBitArray to pass the initial value to memset directly.
This change is being used to pre-populate the prime table with multiples
of 2 by initializing it with the correct sequence of bytes to do so (0xaa)
instead of all 1s. (0xff) This negates the need for the for loop that marks
all even numbers, removing a time-consuming operation from the program.