Count primes afterwards byte by byte
Instead of counting all composite numbers while finding primes, count the
number of primes afterwards byte by byte. This removes a branch and uses a
LUT for the number of active bits in each byte value. Since each byte is
touched only once and no bitmasks are requred, this is a much faster method of
counting.